Modernize analyzer codes for CalibCalorimetry/CastorCalib
