+++ CalibCalorimetry/CastorCalib/interface/CastorLedAnalysis.h
@@ -2,7 +2,6 @@
 #define CastorLedAnalysis_H
 
 #include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/EDAnalyzer.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
