+++ CalibCalorimetry/CastorCalib/plugins/CastorHardcodeCalibrations.cc
@@ -1,10 +1,14 @@
 // -*- C++ -*-
+// ESSource to generate default HCAL/CASTOR calibration objects
+//
 // Original Author:  Fedor Ratnikov
 // Adapted for CASTOR by L. Mundim
 //
 
-#include <memory>
 #include <iostream>
+#include <memory>
+#include <map>
+#include <string>
 
 #include "FWCore/Framework/interface/ValidityInterval.h"
 #include "DataFormats/HcalDetId/interface/HcalCastorDetId.h"
@@ -26,7 +30,35 @@
 #include "Geometry/CaloTopology/interface/HcalTopology.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 
-#include "CastorHardcodeCalibrations.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/Framework/interface/EventSetupRecordIntervalFinder.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+
+#include "CondFormats/CastorObjects/interface/AllObjects.h"
+
+class CastorHardcodeCalibrations : public edm::ESProducer, public edm::EventSetupRecordIntervalFinder {
+public:
+  CastorHardcodeCalibrations(const edm::ParameterSet&);
+  ~CastorHardcodeCalibrations() override;
+
+  void produce(){};
+
+protected:
+  void setIntervalFor(const edm::eventsetup::EventSetupRecordKey&,
+                      const edm::IOVSyncValue&,
+                      edm::ValidityInterval&) override;
+
+  std::unique_ptr<CastorPedestals> producePedestals(const CastorPedestalsRcd& rcd);
+  std::unique_ptr<CastorPedestalWidths> producePedestalWidths(const CastorPedestalWidthsRcd& rcd);
+  std::unique_ptr<CastorGains> produceGains(const CastorGainsRcd& rcd);
+  std::unique_ptr<CastorGainWidths> produceGainWidths(const CastorGainWidthsRcd& rcd);
+  std::unique_ptr<CastorQIEData> produceQIEData(const CastorQIEDataRcd& rcd);
+  std::unique_ptr<CastorChannelQuality> produceChannelQuality(const CastorChannelQualityRcd& rcd);
+  std::unique_ptr<CastorElectronicsMap> produceElectronicsMap(const CastorElectronicsMapRcd& rcd);
+  std::unique_ptr<CastorRecoParams> produceRecoParams(const CastorRecoParamsRcd& rcd);
+  std::unique_ptr<CastorSaturationCorrs> produceSaturationCorrs(const CastorSaturationCorrsRcd& rcd);
+  bool h2mode_;
+};
 
 // class declaration
 //
@@ -229,3 +261,8 @@ std::unique_ptr<CastorSaturationCorrs> CastorHardcodeCalibrations::produceSatura
   }
   return result;
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/SourceFactory.h"
+
+DEFINE_FWK_EVENTSETUP_SOURCE(CastorHardcodeCalibrations);
