+++ CalibCalorimetry/CastorCalib/plugins/CastorPedestalsAnalysis.cc
@@ -5,11 +5,118 @@
 // This code runs 1000x faster and produces all outputs from a single run
 // (ADC, fC in .txt plus an .xml file)
 //
+
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+
+#include "CommonTools/UtilAlgos/interface/TFileService.h"
+#include "DataFormats/Common/interface/Handle.h"
+#include "DataFormats/HcalDetId/interface/HcalGenericDetId.h"
+#include "DataFormats/HcalDetId/interface/HcalElectronicsId.h"
+#include "DataFormats/HcalDetId/interface/HcalDetId.h"
+#include "DataFormats/HcalDigi/interface/HcalDigiCollections.h"
+#include "DataFormats/DetId/interface/DetId.h"
+#include "CondFormats/CastorObjects/interface/CastorPedestals.h"
+#include "CondFormats/CastorObjects/interface/CastorPedestalWidths.h"
+#include "CondFormats/CastorObjects/interface/CastorQIECoder.h"
+#include "CondFormats/CastorObjects/interface/CastorQIEData.h"
+#include "CondFormats/CastorObjects/interface/CastorQIEShape.h"
+#include "CondFormats/CastorObjects/interface/CastorElectronicsMap.h"
+#include "CondFormats/CastorObjects/interface/AllObjects.h"
+
+#include "CalibFormats/CastorObjects/interface/CastorDbRecord.h"
+#include "CalibFormats/CastorObjects/interface/CastorDbService.h"
+#include "CalibFormats/CastorObjects/interface/CastorCalibrations.h"
+#include "CalibFormats/CastorObjects/interface/CastorCalibrationWidths.h"
+
+#include "CalibCalorimetry/CastorCalib/interface/CastorDbASCIIIO.h"
+#include "TBDataFormats/HcalTBObjects/interface/HcalTBTriggerData.h"
+
+#include "TProfile.h"
+#include "TH1.h"
+#include "TH2.h"
+#include "TCanvas.h"
+#include "TStyle.h"
+
+#include <cmath>
+#include <iostream>
+#include <map>
 #include <memory>
-#include "CalibCalorimetry/CastorCalib/interface/CastorPedestalsAnalysis.h"
+#include <iomanip>
+#include <fstream>
+#include <vector>
+#include <string>
+
+namespace edm {
+  class ParameterSet;
+  class Event;
+  class EventSetup;
+}  // namespace edm
+
+struct NewPedBunch {
+  HcalCastorDetId detid;
+  bool usedflag;
+  float cap[4];
+  float capfc[4];
+  float sig[4][4];
+  float sigfc[4][4];
+  float prod[4][4];
+  float prodfc[4][4];
+  int num[4][4];
+};
+
+class CastorPedestalsAnalysis : public edm::one::EDAnalyzer<edm::one::SharedResources> {
+public:
+  //Constructor
+  CastorPedestalsAnalysis(const edm::ParameterSet &ps);
+  //Destructor
+  ~CastorPedestalsAnalysis() override;
+  //Analysis
+  void analyze(const edm::Event &event, const edm::EventSetup &eventSetup) override;
+
+private:
+  //Container for data, 1 per channel
+  std::vector<NewPedBunch> Bunches;
+  //Flag for saving histos
+  bool hiSaveFlag;
+  bool dumpXML;
+  bool verboseflag;
+  int runnum;
+  int firstTS;
+  int lastTS;
+  std::string ROOTfilename;
+  std::string pedsADCfilename;
+  std::string pedsfCfilename;
+  std::string widthsADCfilename;
+  std::string widthsfCfilename;
+  std::string XMLfilename;
+  std::string XMLtag;
+  std::string ZSfilename;
+
+  edm::ESGetToken<CastorDbService, CastorDbRecord> tok_cond_;
+  edm::ESGetToken<CastorElectronicsMap, CastorElectronicsMapRcd> tok_map_;
+
+  TH1F *CASTORMeans;
+  TH1F *CASTORWidths;
+
+  // TH2F *dephist[4];
+  TH2F *dephist;
+
+  bool firsttime;
+
+  edm::InputTag castorDigiCollectionTag;
+};
 
 CastorPedestalsAnalysis::CastorPedestalsAnalysis(const edm::ParameterSet& ps)
     : castorDigiCollectionTag(ps.getParameter<edm::InputTag>("castorDigiCollectionTag")) {
+  usesResource(TFileService::kSharedResource);
+
   hiSaveFlag = ps.getUntrackedParameter<bool>("hiSaveFlag", false);
   dumpXML = ps.getUntrackedParameter<bool>("dumpXML", false);
   verboseflag = ps.getUntrackedParameter<bool>("verbose", false);
@@ -187,6 +294,7 @@ CastorPedestalsAnalysis::~CastorPedestalsAnalysis() {
     //  CastorCondXML::dumpObject (outStream5, runnum, runnum, runnum, XMLtag, 1, (*rawPedsItem), (*rawWidthsItem));
   }
 
+  /*
   if (hiSaveFlag) {
     theFile->Write();
   } else {
@@ -197,6 +305,7 @@ CastorPedestalsAnalysis::~CastorPedestalsAnalysis() {
   }
   theFile->cd();
   dephist->Write();
+  */
   dephist->SetDrawOption("colz");
   dephist->GetXaxis()->SetTitle("module");
   dephist->GetYaxis()->SetTitle("sector");
@@ -246,9 +355,11 @@ CastorPedestalsAnalysis::~CastorPedestalsAnalysis() {
     //dephist[3]->SetDrawOption("colz");
     c2->SaveAs(name2.c_str());
 */
+  /*
   std::cout << "Writing ROOT file... ";
   theFile->Close();
   std::cout << "ROOT file closed.\n";
+  */
 }
 
 // ------------ method called to for each event  ------------
@@ -273,20 +384,23 @@ void CastorPedestalsAnalysis::analyze(const edm::Event& e, const edm::EventSetup
     XMLfilename = runnum_string + "-peds_ADC_complete.xml";
     XMLtag = "Castor_pedestals_" + runnum_string;
 
+    edm::Service<TFileService> fs;
+    fs->mkdir("CASTOR");
+    /*
     theFile = new TFile(ROOTfilename.c_str(), "RECREATE");
     theFile->cd();
     // Create sub-directories
     theFile->mkdir("CASTOR");
     theFile->cd();
+    */
+    CASTORMeans = fs->make<TH1F>("All Ped Means CASTOR", "All Ped Means CASTOR", 100, 0, 9);
+    CASTORWidths = fs->make<TH1F>("All Ped Widths CASTOR", "All Ped Widths CASTOR", 100, 0, 3);
 
-    CASTORMeans = new TH1F("All Ped Means CASTOR", "All Ped Means CASTOR", 100, 0, 9);
-    CASTORWidths = new TH1F("All Ped Widths CASTOR", "All Ped Widths CASTOR", 100, 0, 3);
-
-    dephist = new TH2F("Pedestals (ADC)", "All Castor", 14, 0., 14.5, 16, .5, 16.5);
-    // dephist[0] = new TH2F("Pedestals (ADC)","Depth 1",89, -44, 44, 72, .5, 72.5);
-    // dephist[1] = new TH2F("Pedestals (ADC)","Depth 2",89, -44, 44, 72, .5, 72.5);
-    // dephist[2] = new TH2F("Pedestals (ADC)","Depth 3",89, -44, 44, 72, .5, 72.5);
-    // dephist[3] = new TH2F("Pedestals (ADC)","Depth 4",89, -44, 44, 72, .5, 72.5);
+    dephist = fs->make<TH2F>("Pedestals (ADC)", "All Castor", 14, 0., 14.5, 16, .5, 16.5);
+    // dephist[0] = fs->make<TH2F>("Pedestals (ADC)","Depth 1",89, -44, 44, 72, .5, 72.5);
+    // dephist[1] = fs->make<TH2F>("Pedestals (ADC)","Depth 2",89, -44, 44, 72, .5, 72.5);
+    // dephist[2] = fs->make<TH2F>("Pedestals (ADC)","Depth 3",89, -44, 44, 72, .5, 72.5);
+    // dephist[3] = fs->make<TH2F>("Pedestals (ADC)","Depth 4",89, -44, 44, 72, .5, 72.5);
 
     const CastorElectronicsMap* myRefEMap = &iSetup.getData(tok_map_);
     std::vector<HcalGenericDetId> listEMap = myRefEMap->allPrecisionId();
