+++ CalibCalorimetry/CastorCalib/plugins/CastorTextCalibrations.cc
@@ -1,9 +1,14 @@
+#include <map>
 #include <memory>
 #include <iostream>
 #include <fstream>
+#include <string>
 
 #include "FWCore/Utilities/interface/Exception.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/Framework/interface/EventSetupRecordIntervalFinder.h"
 #include "FWCore/ParameterSet/interface/FileInPath.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
 
 #include "FWCore/Framework/interface/ValidityInterval.h"
 
@@ -27,7 +32,32 @@
 #include "CondFormats/DataRecord/interface/CastorRecoParamsRcd.h"
 #include "CondFormats/DataRecord/interface/CastorSaturationCorrsRcd.h"
 
-#include "CastorTextCalibrations.h"
+
+class CastorTextCalibrations : public edm::ESProducer, public edm::EventSetupRecordIntervalFinder {
+public:
+  CastorTextCalibrations(const edm::ParameterSet&);
+  ~CastorTextCalibrations() override;
+
+  void produce(){};
+
+protected:
+  void setIntervalFor(const edm::eventsetup::EventSetupRecordKey&,
+                      const edm::IOVSyncValue&,
+                      edm::ValidityInterval&) override;
+
+  std::unique_ptr<CastorPedestals> producePedestals(const CastorPedestalsRcd& rcd);
+  std::unique_ptr<CastorPedestalWidths> producePedestalWidths(const CastorPedestalWidthsRcd& rcd);
+  std::unique_ptr<CastorGains> produceGains(const CastorGainsRcd& rcd);
+  std::unique_ptr<CastorGainWidths> produceGainWidths(const CastorGainWidthsRcd& rcd);
+  std::unique_ptr<CastorQIEData> produceQIEData(const CastorQIEDataRcd& rcd);
+  std::unique_ptr<CastorChannelQuality> produceChannelQuality(const CastorChannelQualityRcd& rcd);
+  std::unique_ptr<CastorElectronicsMap> produceElectronicsMap(const CastorElectronicsMapRcd& rcd);
+  std::unique_ptr<CastorRecoParams> produceRecoParams(const CastorRecoParamsRcd& rcd);
+  std::unique_ptr<CastorSaturationCorrs> produceSaturationCorrs(const CastorSaturationCorrsRcd& rcd);
+
+private:
+  std::map<std::string, std::string> mInputs;
+};
 //
 // class decleration
 //
@@ -144,3 +174,8 @@ std::unique_ptr<CastorSaturationCorrs> CastorTextCalibrations::produceSaturation
     const CastorSaturationCorrsRcd& rcd) {
   return produce_impl<CastorSaturationCorrs>(mInputs["SaturationCorrs"]);
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/SourceFactory.h"
+
+DEFINE_FWK_EVENTSETUP_SOURCE(CastorTextCalibrations);
