--- CalibCalorimetry/CastorCalib/plugins/SealModule.cc
@@ -1,8 +1,0 @@
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-#include "FWCore/Framework/interface/SourceFactory.h"
-
-#include "CastorHardcodeCalibrations.h"
-#include "CastorTextCalibrations.h"
-
-DEFINE_FWK_EVENTSETUP_SOURCE(CastorHardcodeCalibrations);
-DEFINE_FWK_EVENTSETUP_SOURCE(CastorTextCalibrations);
